// libs/shared/modules/src/auth-core/models/domain/user.ts
/**
 * Anecdotes API
 * Anecdotes API's
 *
 * OpenAPI spec version: 0.0.3
 * Contact: apiteam@anecdotes.ai
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface User {
  first_name?: string;
  last_name?: string;
  /**
   * The user role
   */
  role?: RoleEnum;
  email?: string;
  last_login?: Date;
  status?: string;
  audit_id?: string;
  audit_frameworks?: string[];
  photo?: string;
  uid?: string;
}

export type RoleEnum = 'admin' | 'user' | 'auditor' | 'it' | 'collaborator' | 'auditors_zone';
export const RoleEnum = {
  Admin: 'admin' as RoleEnum,
  Collaborator: 'collaborator' as RoleEnum,
  User: 'user' as RoleEnum,
  Auditor: 'auditor' as RoleEnum,
  It: 'it' as RoleEnum,
  Zone: 'auditors_zone' as RoleEnum // This is the auditors portal (aka zone) role
};